[Navigation] Fix Nested NavHost animations
There is an issue with Nested NavHost not being animated when their
parents are being navigated away from. This is caused by the Lifecycle
state of the Nested NavHost being below STARTED which is too low for
animation. The visibleEntries flow has access to CREATED entries, so by
using that we can animate the Nested NavHost while its parent still
exists.